ok! works only on horizontal indeed
from the repo https://replicant.fun/tutorials/tic-tac-toe/ see below:

If you’d like a challenge, check out the almost-there tag and try to solve the following tasks:

Highlight winning columns (requires a test for and changes in game.cljc)
Highlight winning diagonals (requires a test for and changes in game.cljc)
Dim the entire game on ties (requires changes in both game.cljc and ui.cljc, specifically the game->ui-data function)
Extra challenge: Add a button to the page to reset the game after it’s over
All these are solved on the main branch in the tutorial repo. Good luck, and have fun!